Finish components for reviewing patch list/details
