+++ src/components/local_patch_detail.jsx
@@ -1,0 +1,14 @@
+"use strict";
+
+var React = require('react')
+
+module.exports = React.createClass({
+  displayName: 'LocalPatchDetail',
+  render: function () {
+    return (
+      <div>
+      <h1>This is the detail for one local patch</h1>
+      </div>
+    )
+  }
+});
