+++ src/components/local_patch_list.jsx
@@ -1,0 +1,14 @@
+"use strict";
+
+var React = require('react')
+
+module.exports = React.createClass({
+  displayName: 'LocalPatchList',
+  render: function () {
+    return (
+      <div>
+      <h1>These are the local patches</h1>
+      </div>
+    )
+  }
+});
